code/event/requirements.txt
requests~=2.28.1
gspread~=5.7.2
qrcode~=7.3.1
Flask~=2.2.2
python-dotenv~=0.21.0
oauth2client~=4.1.3
Pillow~=9.3.0
pytz~=2022.7
line-bot-sdk~=2.4.1
google-api-python-client~=2.70.0
